Tests/LibWeb/Text/expected/HTML/lzw-image-without-EOI-should-still-load.txt

GIF: loaded, width=1, height=1
TIFF: loaded, width=1, height=1
